refactor(builtin): Move logic to separate file
