+++ formal_verification/src/vir_backend/vir_gen/expr_to_vir/expr.rs
@@ -5,6 +5,7 @@ use crate::{
     vir_backend::vir_gen::{
         build_span, build_span_no_id,
         expr_to_vir::{
+            builtins::try_handle_builtin_call,
             expression_location,
             std_functions::{
                 DECREASES, INVARIANT, handle_fv_std_call, loop_decreases_from_call,
@@ -30,7 +31,7 @@ use noirc_frontend::{
     signed_field::SignedField,
 };
 use num_bigint::{BigInt, BigUint};
-use num_traits::{One, Zero};
+use num_traits::One;
 use vir::{
     ast::{
         AirQuant, ArithOp, AutospecUsage, BinaryOp, BinderX, BitwiseOp, CallTarget, CallTargetKind,
@@ -625,7 +626,9 @@ fn ast_call_to_vir_expr(
         return expr;
     }
 
-    if let Some(expr) = try_handle_builtin_call(call_expr, mode, globals) {
+    if let Some(expr) =
+        try_handle_builtin_call(call_expr, mode, globals, ast_expr_to_vir_expr, ast_expr_to_stmt)
+    {
         return expr;
     }
 
@@ -661,145 +664,6 @@ fn ast_call_to_vir_expr(
     };
 
     SpannedTyped::new(&span, &ast_type_to_vir_type(&call_expr.return_type), exprx)
-}
-
-fn try_handle_builtin_call(
-    call_expr: &Call,
-    mode: Mode,
-    globals: &BTreeMap<GlobalId, (String, Type, Expression)>,
-) -> Option<Expr> {
-    let Expression::Ident(function_ident) = call_expr.func.as_ref() else {
-        return None;
-    };
-    let Definition::Builtin(name) = &function_ident.definition else { return None };
-
-    match name.as_str() {
-        "len" | "array_len" => builtin_len(call_expr),
-        "zeroed" => builtin_zeroed(call_expr),
-        "checked_transmute" => builtin_checked_transmute(call_expr, mode, globals),
-        "black_box" => builtin_black_box(call_expr, mode, globals),
-        // Ownership bookkeeping builtins become no-ops in VIR.
-        "array_refcount" | "slice_refcount" | "slice_push_back" | "slice_push_front"
-        | "slice_pop_back" | "slice_pop_front" | "slice_insert" | "slice_remove" => {
-            builtin_unit_noop(call_expr, mode, globals, name)
-        }
-        _ => None,
-    }
-}
-
-fn builtin_len(call_expr: &Call) -> Option<Expr> {
-    assert!(call_expr.arguments.len() == 1, "Expected builtin `len` to receive a single argument");
-
-    let arg_expr = &call_expr.arguments[0];
-    let Some(arg_type) = arg_expr.return_type() else {
-        return None;
-    };
-
-    let array_length = match arg_type.as_ref() {
-        Type::Array(len, _) => *len,
-        _ => return None,
-    };
-
-    let len_bigint = BigInt::from(array_length as i64);
-    let exprx = ExprX::Const(Constant::Int(len_bigint));
-    let span = build_span_no_id(
-        format!("Builtin len call returning {}", array_length),
-        Some(call_expr.location),
-    );
-
-    Some(SpannedTyped::new(&span, &ast_type_to_vir_type(&call_expr.return_type), exprx))
-}
-
-fn builtin_zeroed(call_expr: &Call) -> Option<Expr> {
-    zero_expr_of_type(&call_expr.return_type, Some(call_expr.location))
-}
-
-fn builtin_checked_transmute(
-    call_expr: &Call,
-    mode: Mode,
-    globals: &BTreeMap<GlobalId, (String, Type, Expression)>,
-) -> Option<Expr> {
-    let arg_expr = call_expr.arguments.first()?;
-    Some(ast_expr_to_vir_expr(arg_expr, mode, globals))
-}
-
-fn builtin_black_box(
-    call_expr: &Call,
-    mode: Mode,
-    globals: &BTreeMap<GlobalId, (String, Type, Expression)>,
-) -> Option<Expr> {
-    let arg_expr = call_expr.arguments.first()?;
-    Some(ast_expr_to_vir_expr(arg_expr, mode, globals))
-}
-
-fn builtin_unit_noop(
-    call_expr: &Call,
-    mode: Mode,
-    globals: &BTreeMap<GlobalId, (String, Type, Expression)>,
-    name: &str,
-) -> Option<Expr> {
-    let mut stmts = Vec::new();
-    for arg in &call_expr.arguments {
-        stmts.push(ast_expr_to_stmt(arg, mode, globals));
-    }
-    let span = build_span_no_id(format!("Builtin {} (no-op)", name), Some(call_expr.location));
-    let exprx = ExprX::Block(Arc::new(stmts), None);
-    Some(SpannedTyped::new(&span, &make_unit_vir_type(), exprx))
-}
-
-fn zero_expr_of_type(typ: &Type, location: Option<Location>) -> Option<Expr> {
-    match typ {
-        Type::Field | Type::Integer(..) => {
-            let span = build_span_no_id("Zero literal".to_string(), location);
-            Some(SpannedTyped::new(
-                &span,
-                &ast_type_to_vir_type(typ),
-                ExprX::Const(Constant::Int(BigInt::zero())),
-            ))
-        }
-        Type::Bool => {
-            let span = build_span_no_id("Bool false literal".to_string(), location);
-            Some(SpannedTyped::new(
-                &span,
-                &ast_type_to_vir_type(typ),
-                ExprX::Const(Constant::Bool(false)),
-            ))
-        }
-        Type::Unit => {
-            let span = build_span_no_id("Unit literal".to_string(), location);
-            Some(SpannedTyped::new(
-                &span,
-                &ast_type_to_vir_type(typ),
-                ExprX::Ctor(
-                    Dt::Tuple(0),
-                    Arc::new("tuple%0".to_string()),
-                    Arc::new(Vec::new()),
-                    None,
-                ),
-            ))
-        }
-        Type::Array(len, element_type) => {
-            let mut elements = Vec::new();
-            for _ in 0..*len {
-                elements.push(zero_expr_of_type(element_type, location)?);
-            }
-            let span = build_span_no_id("Array zero literal".to_string(), location);
-            Some(SpannedTyped::new(
-                &span,
-                &ast_type_to_vir_type(typ),
-                ExprX::ArrayLiteral(Arc::new(elements)),
-            ))
-        }
-        Type::Tuple(inner_types) => {
-            let mut fields = Vec::new();
-            for field_type in inner_types {
-                fields.push(zero_expr_of_type(field_type, location)?);
-            }
-            let span = build_span_no_id("Tuple zero literal".to_string(), location);
-            Some(vir::ast_util::mk_tuple(&span, &Arc::new(fields)))
-        }
-        _ => None,
-    }
 }
 
 fn ast_constrain_to_vir_expr(
